+++ app/src/main/java/org/simple/clinic/medicalhistory/MedicalHistoryQuestion.kt
@@ -9,4 +9,5 @@ enum class MedicalHistoryQuestion(@StringRes val questionRes: Int) {
   HAS_HAD_A_STROKE(R.string.medicalhistory_question_stroke),
   HAS_HAD_A_KIDNEY_DISEASE(R.string.medicalhistory_question_kidney),
   DIAGNOSED_WITH_DIABETES(R.string.medicalhistory_question_diabetes),
+  IS_ON_HYPERTENSION_TREATMENT(R.string.medicalhistory_question_is_on_hypertension_treatment)
 }