+++ app/src/main/java/org/simple/clinic/medicalhistory/OngoingMedicalHistoryEntry.kt
@@ -8,6 +8,7 @@ import org.simple.clinic.medicalhistory.MedicalHistoryQuestion.DIAGNOSED_WITH_HY
 import org.simple.clinic.medicalhistory.MedicalHistoryQuestion.HAS_HAD_A_HEART_ATTACK
 import org.simple.clinic.medicalhistory.MedicalHistoryQuestion.HAS_HAD_A_KIDNEY_DISEASE
 import org.simple.clinic.medicalhistory.MedicalHistoryQuestion.HAS_HAD_A_STROKE
+import org.simple.clinic.medicalhistory.MedicalHistoryQuestion.IS_ON_HYPERTENSION_TREATMENT
 
 @Parcelize
 data class OngoingMedicalHistoryEntry(
@@ -15,6 +16,7 @@ data class OngoingMedicalHistoryEntry(
     val hasHadStroke: Answer = Unanswered,
     val hasHadKidneyDisease: Answer = Unanswered,
     val diagnosedWithHypertension: Answer = Unanswered,
+    val isOnHypertensionTreatment: Answer = Unanswered,
     val hasDiabetes: Answer = Unanswered
 ) : Parcelable {
 
@@ -25,6 +27,7 @@ data class OngoingMedicalHistoryEntry(
       HAS_HAD_A_STROKE -> copy(hasHadStroke = answer)
       HAS_HAD_A_KIDNEY_DISEASE -> copy(hasHadKidneyDisease = answer)
       DIAGNOSED_WITH_DIABETES -> copy(hasDiabetes = answer)
+      IS_ON_HYPERTENSION_TREATMENT -> copy(isOnHypertensionTreatment = answer)
     }
   }
 }